+++ src/main/java/com/nhom7/coworkingspace/dto/response/AmenityResponse.java
@@ -1,0 +1,14 @@
+package com.nhom7.coworkingspace.dto.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AmenityResponse {
+
+    private Long id;
+    private String name;
+}